+++ packages/iov-encoding/types/decimal.d.ts
@@ -1,0 +1,14 @@
+/**
+ * A type for arbitrary precision, non-negative decimals.
+ *
+ * Instances of this class are immutable.
+ */
+export declare class Decimal {
+  static fromUserInput(input: string, fractionalDigits: number): Decimal;
+  static fromAtomics(atomics: string, fractionalDigits: number): Decimal;
+  private static verifyFractionalDigits;
+  readonly atomics: string;
+  private readonly data;
+  private constructor();
+  toString(): string;
+}