+++ packages/iov-encoding/types/index.d.ts
@@ -1,4 +1,5 @@
 export { Bech32 } from "./bech32";
+export { Decimal } from "./decimal";
 export { Encoding } from "./encoding";
 export { Int53, Uint32, Uint53, Uint64 } from "./integers";
 export {